+++ src/mapgen/sfc32_test.h
@@ -3,9 +3,9 @@
 // Copyright (c) 2022 by the OlyTag authors.
 // Please see the LICENSE file in the root directory of this repository for further information.
 
-#ifndef OLYTAG_PRNG_TEST_H
-#define OLYTAG_PRNG_TEST_H
+#ifndef OLYTAG_SFC32_TEST_H
+#define OLYTAG_SFC32_TEST_H
 
 int test_prng(void);
 
-#endif //OLYTAG_PRNG_TEST_H
+#endif //OLYTAG_SFC32_TEST_H
